requirements.txt
# Core dependencies
langchain>=0.1.0
langgraph>=0.0.40
langchain-deepseek
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.6.0

# Web framework
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.25.0

# Optional for development
pytest>=7.4.0
pytest-asyncio>=0.21.0
